Switch logic in indexfiles to always remove the uarch flag.
On UFS filesystems, this is unneeded. On ZFS filesystems, the uarch
flag is automatically set on file creation (and maybe update?). This
is causing a lot of churn in the freebsd-update cycle where
discrepancies in the uarch flag makes the system think a lot of metadata
need to be corrected.