Add support for Copy (the intra-server case) to the NFSv4.2 client and server.
The NFSv4.2 server now has a method for VOP_COPY_FILE_RANGE() that will do
the Copy intra-server operation on an NFSv4.2 server. If this operation fails
with EACCES, it will be retried with the other credentials, since an NFS
RPC only takes one set of credentials.
If the operation fails due to lack of support (or both attempts fail with
EACCES), then it falls back on vn_generic_copy_file_range() using read/write.
The NFSv4.2 server uses vn_generic_copy_file_range() to perform the Copy
operation.