Return NFSERR_NOTSUPP for the NFSv4.2 Copy operation if a pNFS server.
A pNFS server would need to proxy the Read/Write operations to the DS(s)
for the NFSv4.2 Copy operation. Normally it should be more efficient for
the client to do the Read/Write operations directly on the DS(s), so
return NFSERR_NOTSUPP for this case to encourage the client to do the I/O
directly on the DS(s).