Add support for _PC_MIN_HOLE_SIZE to the NFSv4.2 client.
Unfortunately there is no NFSv4.2 attribute for this, so all I could come
up with was to try Seek to see if it supported and, if it is, return
mnt_stat.f_iosize as a "best guess".
At least this allows applications to use hole size non-zero to indicate
support for holes.