Factor the code that creates the mbuf chain from nfsvno_read().
This patch factors the code that creates the mbuf chain and associated
iovec from nfsvno_read() into a separate function called nfsrv_createiovec().
This new function is used by nfsvno_getxattr() to avoid code duplication
for this. The patch also slightly re-organizes nfsvno_getxattr() so that
it first finds out the size of the extended attribute before creating
the mbuf chain for VOP_GETEXTATTR().

I need to put the maximum response size for the session in "nd" so
that the size of the attribute can be properly limited.