Add support for the NFSv4.2 Remove Extended Attribute operation.
This also required addditional arguments for nfsrv_checkremove(), so that
"same client" delegations could be avoided.
I think nfsrv_checkremove() should also check for grace before doing
the delegation removal. That would be a fix unrelated to the NFSv4.2
server work that I will look at for a possible commit to head.