Increase the maximum request/response size for NFSv4.2.
Since the Extended Attribute operations may want to send requests/replies
that are larger than the I/O sizes, make the max request/reply as large as
possible when the session is created.
Use these maximum request/response sizes returned in the session reply as
the limits for the Extended Attribute operations.