Move vfs.nfsd.maxcopyrange to vfs.nfs.maxcopyrange so client can use it as well.
This patch makes the NFSv4.2 client use vfs.nfs.maxcopyrange as a limit for
the size of a Copy RPC similar to the NFSv4.2 server.
This limit is intended to make synchronous Copy operations complete in less
than 1 second so that the RPC round trip time remains below 1 second.

I plan on implementing asynchronous Copy for the NFSv4.2 server soon,
since that is what the Linux client uses for large Copies.