Update the nfsstats structure.
I had left slush in the NFSV42_NPROC sized arrays so that I wouldn't have
to do this. But I was a dumbass and didn't leave slush in the NFSV42_NOPS
arrays. I figured that since the NFSv4.2 RFC was published, that there
would be no additional operations defined but RFC-8276 came along and
defined four more of them.
So, the nfsstats structure needed to be updated. I stuck with the
nfsstatv1 name, since it only changed w.r.t. array sizes, which now have
slush for all the arrays.
Hopefully this won't need to be done again.