Add vfs.nfsd.linuxseekdata so that the NFSv4.2 server can be Linux compatible.
Most seem to agree that RFC-7862 states that when a Seek operation with
CONTENT_DATA (SEEK_DATA) is received with an offset == file_size
(or at EOF, if you prefer), that the server should reply NFS_OK with
that offset and eof == true.
However, this breaks the Linux NFSv4.2 client, which expects NFSERR_INVAL
to be replied for this case.
When this sysctl is set non-zero, the server will now reply NFSERR_INVAL
to be Linux compatible. Since the FreeBSD client will handle either reply
correctly and Linux is the only other extant NFSv4.2 client I know of,
this sysctl is enabled by default.