Replace the two sysctls with one to enable a Linux compatible NFSv4.2 server.
To be compatible with the NFSv4.2 client shipping in the 5.n Linux kernel,
two changes in behaviour are currently required. Replace the two sysctls
that enable these with the single sysctl vfs.nfsd.linux42server to minimize
confusion over what needs to be done to be Linux compatible.
There may also be more of these behaviour changes that have not yet been
found during interoperability testing with Linux.