Add a flag for Advise not supported and fix Copy for length == 0.
When an NFSv4.2 reports that the Advise operation is not supported,
note that so that no more Advise operations will be attempted for the mount.
For Copy where the NFSv4.2 server replies a copied length == 0, fix the
client so that it recognizes this as "at EOF".
This will not normally happen, since the client clips the requested
copy length at EOF, however if another client truncates the file, this
could happen for a Linux style NFSv4.2 server.