Fix vnode locking for the case where Seek is proxied to the DS server.
When a Seek is done on a pNFS server, the Seek is proxied to the DS
server. This requires that the vnode remain locked until after the
proxied Seek is completed. This patch fixes the vnode locking for Seek
so that the vnode remains locked until after the proxy call and then
unlocks it just before the VOP_IOCTL() call.