Fix up the NFSv4.2 client/server for pNFS.
This patch consists of assorted fixes needed to make NFSv4.2 work when
the server is a pNFS server. Mostly it consists of fixes so that the client
and layouts use the correct minor version number.
A small fix to the generic newnfs_request() function was needed so that
it properly handles a server reply of NFSERR_MINORVERMISMATCH.
This only matters for NFSv4.2, since it is used to determine which minor
version the DS for a File Layout pNFS server needs.
Also, the patch adds two new sysctls, so that the minor versions of NFSv4
handled by the server can be limited.