Separate errors doing the rpctlscd/rpctlssd upcalls from RPC errors.
Without this patch, the RPC layer errors were being overloaded and
used to indicate failures in the upcalls in the daemons, usually RPC_FAILED.
This patch separates the error returns from the upcalls into a separate
returned value called "reterr".
This cleans up the code, but does not really change any semantics.