Assorted error handling fixups plus a fix for client side handling of non-application data.
The error handling changes more clearly separate upcall errors from
RPC errors.
The cases where the disconnect upcall fails now assumes that the socket
has been closed by a crashed/restarted daemon and does not attempt
to close the socket again.
The ct_rcvstate variable becomes a set of flag bits, so a flag bit
can be set to tell clnt_vc_dotlsupcall() to call clnt_vc_soupcall()
when a socket upcall needed to just return due to a upcall to the
rpctlscd daemon was in progress.