Fix the daemons so that they do not terminate upon SIGPIPE.
There are cases where the other end has closed the TCP connection and
then, if the daemon does an SSL_shutdown(), it gets a SIGPIPE.
Fix the daemons so they just ignore SIGPIPE.
(The code inheritted termination upon SIGPIPE from the gssd, where the
 only socket is the one used for doing upcalls.)

While here, I tweaked the handling of SSL_shutdown() so that the calls
are done only if the value returned via SSL_get_shutdown() indicates
that a call is needed.

It turns out that NFSv3 over TCP is useful for testing the non-application
data records, since the server side krpc shuts down a connection that is
idle for 6minutes. This never happens normally for NFSv4, but does
happen for NFSv3 if you just leave the mount idle for 6minutes.
This generates close alert messages and then a new connection.