Fix a possible bug found by inspection.
If the record is bogus (not likely), the code m_freem()d the record,
but did not set the pointer null. If there was a disconnect of the
socket after this failure, there could be another m_freem() done on
it.
Fixed by setting the pointer null after the m_freem().
Found by inspection of the code.