Thanks to Ben Kaduk's comments, I looked and the 6 minute timeout was actually the NFS server closing the TCP connection.
As such, this patch reduces the timeout to 15sec and does a shutdown(SHUT_RD)
of the socket so that SSL_connect() in the daemon will fail and return an
error.

It does the same thing for the server, although I have not been able to
get the server side daemon stuck in the openssl library. It has been at
least tested by faking a RPC_TIMEDOUT return from the upcall.