Essentially revert changes done to nfsm_uiombuflist() to what is in head/.
The changes allowed the mbuf list to be created using ext_pgs mbufs.
This is used for doing writing to mirrored DSs. However, these writes
require that part of the mbuf list be copied (by nfsm_copym()) and
m_copym() cannot do this for ext_pgs mbufs.
The work to patch m_copym() to do partial copies of ext_pgs mbuf lists
is non-trivial, so I am not doing it at this time.

This does mean that the write requests will be in mbuf clusters and
will need to be copied in the krpc before sosend() for TLS connections,
but will work correctly.