Make sure the NFS server only generates ext_pgs mbufs when there will not be an m_copym() done to cache a reply.
For NFSv3, the caching of a reply is based on the RPC number. For NFSv4.1
and NFSv4.2, the Sequence operation request indicates whether or not to
hold only a copy in the session. Normally, Read, Readdir and Getextattr
will not need a reply copy, but this code makes sure it does not happen
for ext_pgs mbufs.
For NFSv4.0, it cannot use ext_pgs mbufs, since it cannot be sure that
a cached reply copy is not created.