Add code to rpctlscd.c and rpctlssd.c to load "aesni".
This replaces the code in the rc.d scripts, which caused the scripts to
fail if "aesni" could not be loaded.
Loading "aesni" may not be necessary for the cases where TLS has net
interface hardware support, but I do not know how to check for that
and loading aesni.ko seems harmless.