Add support to the rpc.tlsservd daemon for shutting down connections that were verified with no longer valid certificates.
When SIGHUP is posted to rpc.tlsservd, the CRL file is reloaded if one
was specified when the daemon was started.  This patch adds code that
also scans the extant connections after a reload and, for any that
presented a certificate that is no longer valid, shuts the connection down.

This required the code to be changed so that the reload would happen
right away, instead of waiting until the next TLS handshake.