Update the comments w.r.t. the origin of the software.
After an email discussion, there did not seem to be a consensus of what,
if any, changes should/could be done to the copyright.
As such, I have not modified it, but have noted the changes done to
the code in order to use it for NFS-over-TLS.
The copyright is a 2 clause FreeBSD one, so licensing terms are not an
issue.